Implicit locus: Major numerical stability fix (using Long instead of Integer for coefficients, but not for exponents---consider using BigInteger in the future, unsure if it is really needed)
git-svn-id: http://dev.geogebra.org/svn/trunk/geogebra@46095 23ce0884-8a58-47d3-bc5c-ddf1cd5b9f9e

+++ common/src/main/java/org/geogebra/common/kernel/CASGenericInterface.java
@@ -163,7 +163,7 @@ public interface CASGenericInterface extends SettingListener {
 	 * @return the program code
 	 */
 	public String createGroebnerSolvableScript(
-			HashMap<Variable, Integer> substitutions, String polys,
+			HashMap<Variable, Long> substitutions, String polys,
 			String freeVars, String dependantVars, boolean transcext);
 
 	/**
